Support migration from previous discovery server versions
In phase one, the general pool map target is mapped to and from "general"
in the replication protocol, so replication interoperates with discovery
servers without support for mapping the general pool.

In phase two, the pool "general" is mapped to the general pool map target
from the replication protocol, but not to the replication protocol. This
allows interoperation with phase one servers.

In phase three, the replication protocol does not map the general pool target.